requirements.txt
asgiref==3.4.1
Django==3.2.6
django-crispy-forms==1.12.0
Pillow==8.3.1
pkg_resources==0.0.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
opencv-python~=4.5.3.56